Merge more patches from https://github.com/kvm-x86/kvm-unit-tests

See merge request kvm-unit-tests/kvm-unit-tests!85

+++ x86/svm.c
@@ -237,7 +237,7 @@ void svm_setup_vmrun(u64 rip)
 	vmcb->save.rsp = (ulong)(guest_stack + ARRAY_SIZE(guest_stack));
 }
 
-int __svm_vmrun(u64 rip)
+u64 __svm_vmrun(u64 rip)
 {
 	struct guest_regs *regs = this_cpu_guest_regs();
 
@@ -255,7 +255,7 @@ int __svm_vmrun(u64 rip)
 	return (vmcb->control.exit_code);
 }
 
-int svm_vmrun(void)
+u64 svm_vmrun(void)
 {
 	return __svm_vmrun((u64)test_thunk);
 }
